fix: fmt

+++ src/contracts/Lithos.sol
@@ -7,16 +7,16 @@ contract Lithos is ILithos {
     string public constant name = "Lithos";
     string public constant symbol = "LITHOS";
     uint8 public constant decimals = 18;
-    uint public totalSupply = 0;
+    uint256 public totalSupply = 0;
 
-    mapping(address => uint) public balanceOf;
-    mapping(address => mapping(address => uint)) public allowance;
+    mapping(address => uint256) public balanceOf;
+    mapping(address => mapping(address => uint256)) public allowance;
 
     bool public initialMinted;
     address public minter;
 
-    event Transfer(address indexed from, address indexed to, uint value);
-    event Approval(address indexed owner, address indexed spender, uint value);
+    event Transfer(address indexed from, address indexed to, uint256 value);
+    event Approval(address indexed owner, address indexed spender, uint256 value);
 
     constructor() {
         minter = msg.sender;
@@ -36,13 +36,13 @@ contract Lithos is ILithos {
         _mint(_recipient, 50 * 1e6 * 1e18);
     }
 
-    function approve(address _spender, uint _value) external returns (bool) {
+    function approve(address _spender, uint256 _value) external returns (bool) {
         allowance[msg.sender][_spender] = _value;
         emit Approval(msg.sender, _spender, _value);
         return true;
     }
 
-    function _mint(address _to, uint _amount) internal returns (bool) {
+    function _mint(address _to, uint256 _amount) internal returns (bool) {
         totalSupply += _amount;
         unchecked {
             balanceOf[_to] += _amount;
@@ -51,11 +51,7 @@ contract Lithos is ILithos {
         return true;
     }
 
-    function _transfer(
-        address _from,
-        address _to,
-        uint _value
-    ) internal returns (bool) {
+    function _transfer(address _from, address _to, uint256 _value) internal returns (bool) {
         balanceOf[_from] -= _value;
         unchecked {
             balanceOf[_to] += _value;
@@ -64,23 +60,19 @@ contract Lithos is ILithos {
         return true;
     }
 
-    function transfer(address _to, uint _value) external returns (bool) {
+    function transfer(address _to, uint256 _value) external returns (bool) {
         return _transfer(msg.sender, _to, _value);
     }
 
-    function transferFrom(
-        address _from,
-        address _to,
-        uint _value
-    ) external returns (bool) {
-        uint allowed_from = allowance[_from][msg.sender];
-        if (allowed_from != type(uint).max) {
+    function transferFrom(address _from, address _to, uint256 _value) external returns (bool) {
+        uint256 allowed_from = allowance[_from][msg.sender];
+        if (allowed_from != type(uint256).max) {
             allowance[_from][msg.sender] -= _value;
         }
         return _transfer(_from, _to, _value);
     }
 
-    function mint(address account, uint amount) external returns (bool) {
+    function mint(address account, uint256 amount) external returns (bool) {
         require(msg.sender == minter, "not allowed");
         _mint(account, amount);
         return true;
