+++ src/contracts/interfaces/ILithos.sol
@@ -2,11 +2,11 @@
 pragma solidity 0.8.29;
 
 interface ILithos {
-    function totalSupply() external view returns (uint);
-    function balanceOf(address) external view returns (uint);
-    function approve(address spender, uint value) external returns (bool);
-    function transfer(address, uint) external returns (bool);
-    function transferFrom(address, address, uint) external returns (bool);
-    function mint(address, uint) external returns (bool);
+    function totalSupply() external view returns (uint256);
+    function balanceOf(address) external view returns (uint256);
+    function approve(address spender, uint256 value) external returns (bool);
+    function transfer(address, uint256) external returns (bool);
+    function transferFrom(address, address, uint256) external returns (bool);
+    function mint(address, uint256) external returns (bool);
     function minter() external returns (address);
 }
