+++ src/contracts/interfaces/IVeArtProxy.sol
@@ -2,10 +2,8 @@
 pragma solidity 0.8.29;
 
 interface IVeArtProxy {
-    function _tokenURI(
-        uint _tokenId,
-        uint _balanceOf,
-        uint _locked_end,
-        uint _value
-    ) external pure returns (string memory output);
+    function _tokenURI(uint256 _tokenId, uint256 _balanceOf, uint256 _locked_end, uint256 _value)
+        external
+        pure
+        returns (string memory output);
 }
