+++ test/PermissionsRegistry.t.sol
@@ -82,9 +82,7 @@ contract PermissionsRegistryTest is Test {
         assertTrue(registry.hasRole(bytes(GOVERNANCE), user2));
         assertTrue(registry.hasRole(bytes(VOTER_ADMIN), user3));
 
-        address[] memory governanceAddresses = registry.roleToAddresses(
-            GOVERNANCE
-        );
+        address[] memory governanceAddresses = registry.roleToAddresses(GOVERNANCE);
         assertEq(governanceAddresses.length, 2);
 
         vm.stopPrank();
@@ -125,9 +123,7 @@ contract PermissionsRegistryTest is Test {
         assertTrue(registry.hasRole(bytes(GOVERNANCE), user2));
         assertTrue(registry.hasRole(bytes(VOTER_ADMIN), user1));
 
-        address[] memory governanceAddresses = registry.roleToAddresses(
-            GOVERNANCE
-        );
+        address[] memory governanceAddresses = registry.roleToAddresses(GOVERNANCE);
         assertEq(governanceAddresses.length, 1);
         assertEq(governanceAddresses[0], user2);
 
@@ -234,9 +230,7 @@ contract PermissionsRegistryTest is Test {
         assertTrue(registry.hasRole(bytes(GAUGE_ADMIN), user2));
         assertTrue(registry.hasRole(bytes(BRIBE_ADMIN), user3));
 
-        address[] memory governanceAddresses = registry.roleToAddresses(
-            GOVERNANCE
-        );
+        address[] memory governanceAddresses = registry.roleToAddresses(GOVERNANCE);
         assertEq(governanceAddresses.length, 2);
 
         vm.stopPrank();
@@ -376,7 +370,7 @@ contract PermissionsRegistryTest is Test {
 
         string[] memory roles = registry.rolesToString();
         bool found = false;
-        for (uint i = 0; i < roles.length; i++) {
+        for (uint256 i = 0; i < roles.length; i++) {
             if (keccak256(bytes(roles[i])) == keccak256(bytes(NEW_ROLE))) {
                 found = true;
                 break;
@@ -403,10 +397,8 @@ contract PermissionsRegistryTest is Test {
         assertFalse(registry.hasRole(bytes(NEW_ROLE), user1));
 
         string[] memory roles = registry.rolesToString();
-        for (uint i = 0; i < roles.length; i++) {
-            assertTrue(
-                keccak256(bytes(roles[i])) != keccak256(bytes(NEW_ROLE))
-            );
+        for (uint256 i = 0; i < roles.length; i++) {
+            assertTrue(keccak256(bytes(roles[i])) != keccak256(bytes(NEW_ROLE)));
         }
 
         vm.stopPrank();
@@ -426,12 +418,10 @@ contract PermissionsRegistryTest is Test {
         bytes[] memory rolesBytes = registry.roles();
         assertEq(rolesBytes.length, 6);
 
-        uint rolesLength = registry.rolesLength();
+        uint256 rolesLength = registry.rolesLength();
         assertEq(rolesLength, 6);
 
-        address[] memory governanceAddresses = registry.roleToAddresses(
-            GOVERNANCE
-        );
+        address[] memory governanceAddresses = registry.roleToAddresses(GOVERNANCE);
         assertEq(governanceAddresses.length, 1);
         assertEq(governanceAddresses[0], user1);
